+++ openrouteservice/src/main/java/heigit/ors/isochrones/statistics/postgresql/PostgresSQLStatisticsProvider.java
@@ -111,68 +111,78 @@ public class PostgresSQLStatisticsProvider extends AbstractStatisticsProvider {
      * @throws Exception
      */
     @Override
-    public void close() throws Exception {
+    public void close() {
         if (_dataSource != null) {
             _dataSource.close();
             _dataSource = null;
         }
     }
 
+    /**
+     * The function takes an {@link Isochrone} as an input along with a {@link String}[] holding the attributes parameters set in the api attributes variable.
+     * For now only pop_area and pop_total can be asked. Together or as single values.
+     *
+     * @param isochrone  {@link Isochrone} as input.
+     * @param properties {@link String}[] as input holding the attributes parameters.
+     * @return Returns a double[] holding the desired values in the order that was asked for in the attributes.
+     * @throws Exception If the query doesn't return any values or the sql is corrupt, an {@link Exception} will be thrown.
+     */
     @Override
     public double[] getStatistics(Isochrone isochrone, String[] properties) throws Exception {
         int nProperties = properties.length;
         double[] res = new double[nProperties];
-
         Connection connection = null;
-        PreparedStatement statement = null;
-        Exception exception = null;
-
+        PreparedStatement preparedStatement = null;
+        ResultSet resultSet;
         try {
-            String strParams = "";
-            String strRatioParams = "";
-            for (int i = 0; i < nProperties; i++) {
-                strRatioParams += "SUM(overlap_ratio * \"" + properties[i] + "\")";
-                strParams += "\"" + properties[i] + "\"";
-                if (i < nProperties - 1) {
-                    strRatioParams += ", ";
-                    strParams += ", ";
+            String sql = null;
+            for (String property : properties) {
+                String polyGeom = isochrone.getGeometry().toText();
+                switch (property) {
+                    case "area":
+                    case "pop_total":
+                        sql = "SELECT ROUND(ST_Area(poly)/1000000) as area, ROUND(SUM((ST_SummaryStats(ST_Clip(" + _geomColumn + ",poly))).sum)) as pop_total FROM " + _tableName + ", ST_Transform(ST_GeomFromText( '" + polyGeom + "', 4326), 954009) as poly WHERE ST_Intersects(poly," + _geomColumn + ")GROUP BY poly;";
+                        break;
+                    default:
+                        break;
                 }
+
             }
-
-            String sql = "SELECT " + strRatioParams + " FROM " +
-                    "(" +
-                    "SELECT" + strParams + "," + String.format("ST_Area(ST_Intersection(%s, ?)) / ST_Area(%s) overlap_ratio FROM %s WHERE ST_Intersects(%s, ?)", _geomColumn, _geomColumn, _tableName, _geomColumn) +
-                    ") AS tbl";
-
-            byte[] geomBytes = geometryToWKB(isochrone.getGeometry());
-
             connection = _dataSource.getConnection();
             connection.setAutoCommit(false);
+            preparedStatement = connection.prepareStatement(sql);
 
-            statement = connection.prepareStatement(sql);
-            statement.setBytes(1, geomBytes);
-            statement.setBytes(2, geomBytes);
-
-            ResultSet resSet = statement.executeQuery();
-
-            if (resSet.next()) {
-                for (int i = 0; i < nProperties; i++)
-                    res[i] = resSet.getDouble(i + 1);
+            resultSet = preparedStatement.executeQuery();
+            // check if the resultSet contains values
+            if (resultSet.next()) {
+                // check for each property if the result contains a value
+                int propertyCounter = 0;
+                for (String property : properties) {
+                    int i = 0;
+                    // check each value
+                    while (i < nProperties) {
+                        // Get the column name
+                        String columnName = resultSet.getMetaData().getColumnName(i + 1);
+                        // If a value fits the current property it is set in the correct place in the result[]
+                        if (columnName.equals(property))
+                            res[propertyCounter] = resultSet.getDouble(i + 1);
+                        i++;
+                    }
+                    propertyCounter++;
+                }
             }
         } catch (Exception ex) {
             LOGGER.error(ex);
-            exception = new InternalServerException(LocationsErrorCodes.UNKNOWN, "Unable to retrieve data from the data source.");
+            throw new InternalServerException(LocationsErrorCodes.UNKNOWN, "Unable to retrieve data from the data source.");
         } finally {
-            if (statement != null)
-                statement.close();
+            if (preparedStatement != null) {
+                preparedStatement.close();
+            }
 
-            if (connection != null)
+            if (connection != null) {
                 connection.close();
+            }
         }
-
-        if (exception != null)
-            throw exception;
-
         return res;
 
     }
@@ -236,6 +246,15 @@ public class PostgresSQLStatisticsProvider extends AbstractStatisticsProvider {
 
     }
 
+    /**
+     * This function translates a {@link Geometry} object into a byte[] representation.
+     * It was used by getStatisticsOld().
+     *
+     * @param geom {@link Geometry} object holding the Geometry (e.g. LINESTRING...).
+     * @return byte[] holding the byte representation of the {@link Geometry} object.
+     * @throws IOException
+     */
+    @Deprecated
     private byte[] geometryToWKB(Geometry geom) throws IOException {
         WKBWriter wkbWriter = new WKBWriter();
         ByteArrayOutputStream bytesStream = new ByteArrayOutputStream();
@@ -246,6 +265,11 @@ public class PostgresSQLStatisticsProvider extends AbstractStatisticsProvider {
         return geomBytes;
     }
 
+    /**
+     * Returns the driver name to be queried against the {@link heigit.ors.isochrones.statistics.StatisticsProviderFactory}.
+     *
+     * @return Returns the provider name in a {@link String} format.
+     */
     @Override
     public String getName() {
         return "postgresql";